feat: refactor `AlbumShare` functionality and UI components
